Fixed the issue with the usuario table
The problem was that in src/main/java/Modelo/Dao/UsuarioDAO.java the
table usuario, was referenced as Usuario (starting with uppercase), so
the the response from the driver was that the tables ollitape.Usuario
did not exist. This fix should remove that message and the user list
should appear normally now.

	modified:   src/main/java/Modelo/Dao/UsuarioDAO.java