Updated the requirements for the project
Now, I am using the lastest Django LTS release (4.2) and have updated
the djangorestframework and drf-spectacular, to their latest versions.
But, as of right now, I am yet to test, that they can actually work with
Django 4.2, drf only states it is compatible with Django 4.1. I might
need to revert to an earlier Django version if needed.