Changes the way AppPath builds the private paths
Now, instead of using the Path class, I just used string concatenation
using the usual / character as the path separator. This was done in
hopes of fixing a Linux/Windows path incompatibility problem. Before,
only the first character ( in '/WEB-INF') was a slash (/), and the rest
of separators were given by the Path class (system dependent). But, that
was the state in which the error presented itself.

The application is still working fine in Linux after the changes, but
whether it fixes the error in Windows, that is still to be seen.