Functional Kubernetes configuration
All Kubenetes entities are created correctly, the only problem is, the
'Administrador' type users for the mainapp, there is none of that type
created in the login microservice by default. So I had to manually run:

kubectl exec -it logindb-deployment-59c855dd98-lgdxl  -- psql devdb
mainapp

To update a user I created using the signup endpoint, and change its
user_type to 'Administrator'. After that, everything seems to be working
fine.